// src/api/openapi/model/articleUpdateV2.ts
/**
 * FastAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ArticleUpdateV2 { 
    favorite: boolean;
    modNumber: string;
    nameDE: string;
    nameIT: string;
    price: number;
    unitID: number;
}